nodes/node_1/requirements.txt
flask
bson
pymongo==3.7.2
requests
kafka-python